+++ tests/backends/test_macOS.py
@@ -3,7 +3,7 @@ import sys
 import pytest
 
 from keyring.testing.backend import BackendBasicTests
-from keyring.backends import OS_X
+from keyring.backends import macOS
 
 
 def is_osx_keychain_supported():
@@ -13,4 +13,4 @@ def is_osx_keychain_supported():
 @pytest.mark.skipif(not is_osx_keychain_supported(), reason="Needs macOS")
 class TestOSXKeychain(BackendBasicTests):
     def init_keyring(self):
-        return OS_X.Keyring()
+        return macOS.Keyring()